// src/app/sources/sheet-report-list/share/model/sheet-report-list-item.ts
export class SheetReportListItem {
    id?: number;
    researchSheetId?: string;
    type?: number;
    createdDate?: number;
    createdBy?: string;
    content?: string;
    likeCount?: number;
    disLikeCount?: number;
    fileId?: string;
    link?: string;
    commentId?: string;
    status?: boolean;
    changeStatusBy?: string;
    description?: string;
    resourceTitle?: string;
    sheetTitle?: string;
}